@@ -3,33 +3,32 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@chakra-ui/icons": "^1.1.7",
-    "@chakra-ui/react": "^1.8.8",
+    "@chakra-ui/icons": "^2.0.1",
+    "@chakra-ui/react": "^2.1.2",
     "@emotion/react": "^11.9.0",
     "@emotion/styled": "^11.8.1",
-    "@fontsource/cabin": "^4.5.5",
-    "@fontsource/inter": "^4.5.7",
-    "@fontsource/noto-sans-jp": "^4.5.7",
-    "@fontsource/poppins": "^4.5.5",
+    "@fontsource/cabin": "^4.5.6",
+    "@fontsource/inter": "^4.5.11",
+    "@fontsource/noto-sans-jp": "^4.5.10",
+    "@fontsource/poppins": "^4.5.8",
     "@fontsource/roboto": "^4.5.7",
     "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^12.1.4",
-    "@testing-library/user-event": "^13.5.0",
+    "@testing-library/react": "^13.3.0",
+    "@testing-library/user-event": "^14.2.0",
     "animate.css": "^4.1.1",
     "axios": "^0.27.2",
     "dotenv": "^16.0.1",
-    "framer-motion": "^6.2.9",
-    "gsap": "^3.10.3",
-    "html-react-parser": "^1.4.12",
-    "react": "^17.0.2",
+    "framer-motion": "^6.3.10",
+    "gsap": "^3.10.4",
+    "html-react-parser": "^1.4.14",
+    "react": "^18.1.0",
     "react-anchor-link-smooth-scroll": "^1.0.12",
-    "react-dom": "^17.0.2",
-    "react-intersection-observer": "^9.2.0",
+    "react-dom": "^18.1.0",
+    "react-intersection-observer": "^9.2.2",
     "react-router-dom": "^6.3.0",
-    "react-scripts": "5.0.0",
+    "react-scripts": "5.0.1",
     "react-slick": "^0.29.0",
     "slick-carousel": "^1.8.1",
-    "swiper": "^8.1.4",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
