@@ -1,5 +1,6 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import { StrictMode } from "react";
+import { createRoot } from "react-dom/client";
 import "./index.css";
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
@@ -13,15 +14,17 @@ import "@fontsource/poppins";
 import "@fontsource/poppins/700.css";
 import "@fontsource/roboto";
 
-ReactDOM.render(
-  <React.StrictMode>
-    <BrowserRouter>
-      <ChakraProvider theme={Theme}>
-        <App />
-      </ChakraProvider>
-    </BrowserRouter>
-  </React.StrictMode>,
-  document.getElementById("root")
+const container = document.getElementById("root");
+const root = createRoot(container);
+
+root.render(
+  //<StrictMode>
+  <BrowserRouter>
+    <ChakraProvider theme={Theme}>
+      <App />
+    </ChakraProvider>
+  </BrowserRouter>
+  //</StrictMode>
 );
 
 // If you want to start measuring performance in your app, pass a function
